+++ src/lib.rs
@@ -119,7 +119,9 @@ pub use crate::algorithms::properties::eigenvector::eigenvector_centrality;
 pub use crate::algorithms::properties::girth::girth;
 pub use crate::algorithms::properties::harmonic::harmonic_centrality;
 pub use crate::algorithms::properties::harmonic_weighted::harmonic_centrality_weighted;
-pub use crate::algorithms::properties::hits::{HitsScores, hub_and_authority_scores};
+pub use crate::algorithms::properties::hits::{
+    HitsScores, hub_and_authority_scores, hub_and_authority_scores_weighted,
+};
 pub use crate::algorithms::properties::is_acyclic::is_acyclic;
 pub use crate::algorithms::properties::is_complete::is_complete;
 pub use crate::algorithms::properties::is_dag::is_dag;